Update azurite
We need to change the name of the service from azurite.local, to
azurite. Problem is that they introduced this breaking change

https://github.com/Azure/Azurite/pull/1043/files

It looks like a simple fix, just move the container name from path
to subdomain. But then it simply do not work, because the domain
<container>.azurite.local cannot be resolved. There is probably a
way to do it in the dns config of the cluster, but that is outside
of scope what we want to do. So I just renamed it. Other solution
will be to add static ip and than use that in the connection string.