Finally make it work
AKS and scaleset both use system assigned managed identity, so there
is no need to create a separate service principal for them. This works
out great, but it create a lot of orphaned service principals in the
acr. To clean them up, I have added a powershell script that will
delete all role assignments that use service principals that are not
used.  It is not perfect, but it gets the job done and it seems like
there is no other way to do it. At least not an easy one.